define gráfico 57 como ambulancia financiada pelo projeto
